feat: finish auth & cart

+++ src/cart/cart.service.ts
@@ -1,5 +1,5 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
-import { CreateCartDto, ITotalCartSum, RemoveCartDto } from './dto/cart.dto'
+import { CreateCartDto, ITotalCartSum, PlusOrMinus, RemoveCartDto } from './dto/cart.dto'
 import { PrismaService } from 'src/prisma/prisma.service';
 
 @Injectable()
@@ -95,26 +95,31 @@ export class CartService {
         return updateUserCart
 
     }
-
-    async plusOrMinus(data: CreateCartDto) {
-        if (data.type !== 'plus' && data.type !== 'minus') throw new BadRequestException('Invalid type')
-        const cart = await this.prisma.cart.findFirst({
-            where: {
-                userId: data.userId
-            },
+    async getCart(userId: string) {
+        return await this.prisma.cart.findFirst({
+            where: { userId },
             include: {
-                cartItems: true
+                cartItems: {
+                    include: {
+                        productItemInfo: {
+                            include: {
+                                product: true,
+                                memory: true
+                            }
+                        }
+                    }
+                }
             }
         })
-        const validCartItem = cart.cartItems.some(({ productItemInfoId }) => productItemInfoId !== data.productInfoId)
-        if (validCartItem) throw new BadRequestException('No cart Item')
-
+    }
+    async plusOrMinus(data: PlusOrMinus) {
+        if (data.type !== 'plus' && data.type !== 'minus') throw new BadRequestException('Invalid type')
         const cartItem = await this.prisma.cartItem.findFirst({
             where: { id: data.cartItemId }
         })
 
         if (!cartItem) throw new BadRequestException('No cartItem')
-        if (cartItem.quantity === 1) {
+        if (data.type === 'minus' && cartItem.quantity === 1) {
             await this.prisma.cartItem.delete({
                 where: { id: cartItem.id }
             })
